requirements.txt
openai
fastapi
uvicorn
python-dotenv
langfuse
langchain
langchain_openai